refactored abstract data accessclass to the utils folder
