+++ backend/utils/firebase_data_access_implementation.py
@@ -1,7 +1,6 @@
 from utils.firebase import db
-from .AbstractDataAccess import AbstractDataAccess
-
-class Firebase_data_access(AbstractDataAccess):
+from .abstract_data_access import AbstractDataAccess
+class FirebaseDataAccess(AbstractDataAccess):
     def get_table_from_database(table_to_access: str) -> dict:
         """
         return a dict mapping the key of the firestore collection to the data within its rows
